Moved stair blocks into one class and created StairsBlock parent class. Close #33
